修改用户表bug，同时处理UnifyResponse的CreateSuccess等四个方法

+++ src/main/java/com/xidian/qunzhi/pojo/User.java
@@ -13,6 +13,7 @@ public class User{
 
     private String nickname;
 
+
     private String realname;
 
     private String password;
@@ -30,7 +31,7 @@ public class User{
     private String district;
 
     @Column(name = "is_admin")
-    private Integer isAdmin;
+    private Short isAdmin;
 
 
     @JsonIgnore
